Add gfxsubs.bas, moving zooming/smoothing from gfx_fb and gfx_sdl there. No change to gfx_sdl brokenness.
Also, gfx_alleg compiles again. I'm surprised that it still works

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2674 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ allmodex.bas
@@ -139,7 +139,7 @@ sub setmodex()
 	dim i as integer
 
 	'initialise software gfx
-	for i = 0 to 3
+	for i as integer = 0 to 3
 		spage(i) = callocate(320 * 200)
 	next
 	'other spage slots are for temporary pages
@@ -179,7 +179,7 @@ sub restoremode()
 	mutexdestroy keybdmutex
 
 	'clear up software gfx
-	for i = 0 to ubound(spage)
+	for i as integer = 0 to ubound(spage)
 		if spage(i) then
 			deallocate(spage(i))
 			spage(i) = NULL
